add unit test for general modulus
the most general modulus possible such that the multiplicative group is cyclic is p^k or 2*p^k. in this case, we can now compute the NTT. this provides a unit test for this behavior. note that n, the polynomial size, must be a power of 2. since \phi(p^k) = p^k(p-1), this means we need p-1 to contain sufficiently many powers of 2.